backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "google-generativeai>=0.8.5",
    "litellm>=1.72.6",
    "ollama>=0.5.1",
    "pydantic[email]>=2.11.7",
    "pydantic-ai>=0.2.19",
    "python-dotenv>=1.1.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "uvicorn>=0.34.3",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "asyncpg>=0.29.0",
    "aiosqlite>=0.20.0",
    "pydantic-settings>=2.0.0",
    "greenlet>=3.0.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=4.0.0,<5.0.0",
    "authlib>=1.6.0",
    "httpx>=0.28.1",
    "psycopg2-binary>=2.9.10",
    "cryptography>=42.0.0",
    "pydub>=0.25.1",
    "ffmpeg-python>=0.2.0",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "aiofiles>=24.1.0",
    "boto3>=1.38.37",
    "mcp>=1.9.4",
    "aiohttp>=3.12.13",
    "docker>=7.0.0",
    "websockets>=15.0.1",
    "watchdog>=6.0.0",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
    "pexpect>=4.9.0",
    "python-socketio==5.8.0",
    "python-engineio==4.7.1",
    "aiodocker>=0.24.0",
]

[tool.uv]
dev-dependencies = []

[tool.flake8]
ignore = ["E501"]
